Algoritmo de busqueda de eventos mejorado.
Cambios realizados en el modelo para la adaptación:

	-Artista {
		+public boolean contiene(String pFiltro)
	}

	-Disco{
		+public void addCancion(int pNum, int pDur, int pPunt, String pTitulo,
String pComp, Genero pGen)
	}

	-Entrada{
		+public int getPrecio()
	}

	-EventoMusical{
		+public EventoMusical(int pMaxEntradas, String pNombre, String pLugar,
Date pFecha, Date pHora, int pPrecioEntrada, int pPrecioVip, Genero
pGenero, ListaGrupo pGruposParticipantes, ListaEntrada pListaEntradas)
==> Se quita el id de la constructora.	
		+public int cuantasTiene()
	}

	-Filtrar{
		+public Filtrar(String pLista)
		+public String getLista()
		+public Boolean fn(Artista ar)
	}

	-Grupo{
		+public Artista buscarComponente(String pComponente)
		+public Disco buscarDisco(String pNombre)
		+public void anadirDisco(String pNom, Date pAno, Genero pGen, int
pDur)
		+public Iterator<String> actArt()
		+public Iterator<String> actDis()
		+public Iterator<String> actArtAnt()
		+public boolean contiene(String pFilter)
	}

	-ListaArtista{
		+public void addArtista(String pNombre, Tipo pTipo)
		+public Artista buscarArtista(String pNombre)
		+public ArrayList<Artista> getArtistas()
		+public Iterator<String> rellenar()
	}

	-ListaCancion{
		+public Iterator<String> actualizar()
	}

	-ListaDisco{
		+public Iterator<String> actualizar()
	}

	-ListaEntrada{
		+public int calcularBeneficio() ==> Se cambia float por int
	}

	-ListaEventoMusical{
		+public EventoMusical buscarEvento(String pNom)
		+public int obtenerBeneficio() ==> Se cambia float por int
		+public int obtenerBeneficio(EventoMusical pEvento) ==> Se cambia
float por int
		+public int obtenerPrecioEntrada(EventoMusical pEvento) ==> Se cambia
float por int
		+public int obtenerPrecioEntradaVip(EventoMusical pEvento) ==> Se
cambia float por int
		+public Iterator<String> rellenar()
		+public Iterator<String> actualizar(String pFilter)
		+private boolean buscar(String pFilter, String pComp)
	}

	-ListaGrupo{
		+public void addGrupo(String pNombre, String pLogo)
		+public Grupo getGrupo(String pNombre)
		+public void eliminarGrupo(String pNombre)
		+public Iterator<String> rellenar()
	}

	-Ordenar{
		+public Ordenar()
		+public int compare(String s1, String s2)
	}

	-Tipo{
		VOZ,GUITARRA,BAJO,BATERIA,PIANO,DJ,SOLISTA
	}

	-CatalogoGrupoArtista{
		+public Grupo getGrupo(String pNombre)
		+public void addGrupo(String pNombre, String pLogo)
		+public void anadirIntegrante(Artista pArtista, Grupo pGrupo)
		+public void reemplazarIntegrante(Artista pArtistaN, Artista
pArtistaV, String pGrupo)
		+public Disco buscarDisco(String pNombre, Grupo pGrupo)
		+public int obtenerBeneficio()
		+public int obtenerBeneficio(EventoMusical pEvento)
		+public void anadirDisco(Grupo pGrupo, String pNom, Date pAno, Genero
pGen, int pDur)
		+public Iterator<String> ordenar()
		+public Iterator<String> filtrarArtistas(String pFilter)
		+public Iterator<String> filtrarGrupos(String pFilter)
		+public Iterator<String> rellenar()
		+public Iterator<String> rellenarArtistas()
	}

	-CatalogoEventoMusical{
		+public EventoMusical buscarEvento(String pNom)
		+public int obtenerBeneficio(EventoMusical pEvento)
		+public int obtenerBeneficio()
		+public int obtenerPrecioEntrada(EventoMusical pEvento)
		+public int obtenerPrecioEntradaVip(EventoMusical pEvento)
		+public Iterator<String> rellenar()
		+public Iterator<String> actualizar(String pFilter)
		+public void anadirConcierto(Concierto pCon)
		+public void anadirFestival(Festival pFes)
		+public void anadirRPrensa(RuedaPrensa pRP)
	}

+++ SoundBlast/src/packMae/CatalogoGrupoArtista.java
@@ -96,11 +96,11 @@ public class CatalogoGrupoArtista {
 		return CatalogoEventoMusical.getCatalogoEventoMusical().buscarFestivalesPorGenero(pGenero);
 	}
 	
-	public float obtenerBeneficio() {
+	public int obtenerBeneficio() {
 		return CatalogoEventoMusical.getCatalogoEventoMusical().obtenerBeneficio();
 	}
 	
-	public float obtenerBeneficio(EventoMusical pEvento) {
+	public int obtenerBeneficio(EventoMusical pEvento) {
 		return CatalogoEventoMusical.getCatalogoEventoMusical().obtenerBeneficio(pEvento);
 	}
